refactor(chats): extract NoActiveChatState, HecateTaskApprovalsBanner, ChatSettingsPanel
ChatView.tsx had three internal subcomponent clusters that were
self-contained and ripe for extraction:

- NoActiveChatState: 25 LOC empty-state for when no chat is selected.
- HecateTaskApprovalsBanner + 8 helpers (pendingHecateTaskApprovals,
  activeTaskBackedHecateSegment, etc.): the in-chat approval-prompt
  cluster. ChatView still calls the two pure helpers; the rest are
  internal to the banner module.
- ChatSettingsPanel + 7 subcomponents + 2 format helpers
  (formatAgentContextUsage, formatAgentReportedCost): the
  right-side settings sheet.

ChatView.tsx: 1560 -> 858 LOC. New files:
- ChatNoActiveState.tsx (26 LOC)
- HecateTaskApprovalsBanner.tsx (228 LOC)
- ChatSettingsPanel.tsx (454 LOC)

No UX change; no prop-shape change; HecateTaskApproval type is now
exported from HecateTaskApprovalsBanner.tsx.
